+++ packages/form/packages/bond_form_bloc/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
-## 0.0.1+1
-* Updated bond_form dependency to ^0.0.22+5
+## 0.0.1+2
+* Updated bond_form dependency to ^0.0.22+6
 
 ## 0.0.1
 - Initial release of the `bond_form_bloc` package.